fixed a drawing issue, tried to fix scaling issue
Former-commit-id: 65532ab359111fdfb914f38489682dcc784f383b

+++ src/main/java/com/compomics/pepshell/ProgramVariables.java
@@ -17,7 +17,7 @@ public class ProgramVariables {
     public static Color PROTEINCOLOR = new Color(93, 100, 202);
     public static Color PEPTIDECOLOR = new Color(223, 156, 55);
     public static Color DOMAINCOLOR = Color.GREEN;
-    public static double SCALE = 1;
+    public static int SCALE = 1;
     public static int VERTICALSIZE = 15;
     public static boolean RESIZEPANELS = true;
     public static File EXPORTFOLDER = new File("C:/Users/Davy");
